subscribe_ml/requirements.txt
quixstreams
python-dotenv
pandas
influxdb_client